extract simulation into lib
