--- confero-front/src/components/connections/graph.ts
@@ -1,34 +1,0 @@
-import d3 from "d3";
-
-// import candidates from "../../data/candidates.json";
-// import connections from "../../data/connections.json";
-
-// const main = async () => {
-//   const svg = d3.select("svg");
-
-//   const width = 960;
-//   const height = 600;
-
-/* Force Simulation */
-
-// $("#candidates").val("");
-// $("#candidates").autocomplete({
-//   source: candidates.map(candidate => ({
-//     label: candidate.name,
-//     id: candidate.id
-//   })),
-//   select: (event, ui) => updateHighlight(ui.item.id)
-// });
-
-// node.on("click", ({ id }) => updateHighlight(id));
-
-// $("#clear").click(() => {
-//   $("#candidates").val("");
-
-//   node
-//     .style("opacity", 1)
-//     .attr("stroke", "white")
-//     .attr("stroke-width", 1.5);
-//   link.style("opacity", 1).attr("stroke", "#999");
-// });
-// };
